requirements.txt
psycopg2-binary==2.9.10
requests==2.32.3
bs4
flask
indic-nlp-library